requirements.txt
asyncio
aiofiles
loguru
python-dotenv
retry
openpyxl
urllib3
PyExecJS
requests
beautifulsoup4
protobuf
protobuf-to-dict
argparse
websockets
websocket-client
py-mini-racer
